Split fixtures into multiple files
