+++ packages/crypto-messages/source/serializer.test.ts
@@ -19,7 +19,7 @@ import {
 	serializedProposalDataWithValidRound,
 	serializedProposalDataWithValidRoundForSignature,
 	serializedProposalForSignature,
-} from "../test/fixtures/proposal";
+} from "../test/fixtures/index.js";
 import { prepareSandbox } from "../test/helpers/prepare-sandbox";
 import { Deserializer } from "./deserializer";
 import { Serializer } from "./serializer";
