+++ packages/crypto-messages/test/fixtures/index.ts
@@ -1,0 +1,4 @@
+export * from "./block.js";
+export * from "./precommit.js";
+export * from "./prevote.js";
+export * from "./proposal.js";
