+++ packages/crypto-messages/test/fixtures/proposal.ts
@@ -1,68 +1,6 @@
-import { Enums } from "@mainsail/constants";
 import type { Contracts } from "@mainsail/contracts";
-import { BigNumber } from "@mainsail/utils";
 
-export const validatorMnemonic =
-	"sudden head royal retire duck discover danger then basic rice wish left whip chronic enrich sun behind idea remind retire coyote select goddess exile";
-
-export const blockHeader: Contracts.Crypto.BlockHeader = {
-	fee: BigNumber.make("10000000000"),
-	gasUsed: 2000,
-	hash: "82139a7708157c8e2b78f0db38216924c8a17f82e77d5997fb280b1435a6cc97",
-	logsBloom:
-		"00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-	number: 2,
-	parentHash: "0000000000000000000000000000000000000000000000000000000000000000",
-	payloadSize: 226,
-	proposer: "0xB559F4FbB75c378CDd3Dd7CcbFeff9c5c2094E55",
-	reward: BigNumber.ZERO,
-	round: 1,
-	stateRoot: "0000000000000000000000000000000000000000000000000000000000000000",
-	timestamp: 1703128709748,
-	transactionsCount: 2,
-	transactionsRoot: "f01a3a2a2990990a64211feb47e2fa25c048decb3420ee52562fdc4931225c0f",
-	version: 1,
-}
-
-export const blockData: Contracts.Crypto.BlockData = {
-	...blockHeader,
-	transactions: [
-		{
-			data: "",
-			from: "0xE33074cBE63A1f86cc8EAb97b2099732F15284fE",
-			gasLimit: 1_000_000,
-			gasPrice: 5_000_000_000,
-			hash: "65d43d2bed464b6bb1df8d6c0136316d3a3559569904fb16d5e3a8d71ebc2ebc",
-			network: 10_000,
-			nonce: BigNumber.ZERO,
-			r: "921101a4583fb153ec00e501f3c2e2636114e1c8c58d2df8a19426cc066a6768",
-			s: "22db4bce1e0ace485ce0838d178b4d5bcfa9f69b315a14c580d9b01e5c980bdd",
-			senderLegacyAddress: "DUQKzkR4BP5UcWayJMpRXtpEXuw4zPbWhe",
-			senderPublicKey: "02daa6f404dbd49c9c74a3d88d65b967a9b51d3465c92833e8e2ede11e7242f014",
-			to: "0xBe89811e15f611C1db12e59679b6F3DC1F430155",
-			transactionIndex: 0,
-			v: 0,
-			value: BigNumber.ZERO,
-		},
-		{
-			data: "",
-			from: "0x8ae872F64bA0731f66F847f4e8Fc0796dF0bCc08",
-			gasLimit: 1_000_000,
-			gasPrice: 5_000_000_000,
-			hash: "7a74379424e07173c137dd89c29f497b35d8013815b03096b51e67f87c453574",
-			network: 10_000,
-			nonce: BigNumber.ONE,
-			r: "6c9842bc78c2f68468cbf8a8f3fec0ae2679707ffb606a4b373dd01a02af55fc",
-			s: "1a4c4d984d750678fb204ce8b7e97d860c974ac1a423f098dc4921acd2be0c7d",
-			senderLegacyAddress: "DQJTK7of6bPUfJEuL9gUV4qnUyq72eskKe",
-			senderPublicKey: "03043bfdf530d59e919323a33d0c8f5ca43f6b50dfe753c9b3e987a4a5233a2a15",
-			to: "0xBe89811e15f611C1db12e59679b6F3DC1F430155",
-			transactionIndex: 1,
-			v: 0,
-			value: BigNumber.ZERO,
-		},
-	]
-};
+import { blockHeader } from "./block.js";
 
 export const serializedBlock =
 	"000174ba618a8c01020000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200d007000000000000000000000000000000000000000000000000000000000002540be4000000000000000000000000000000000000000000000000000000000000000000e2000000f01a3a2a2990990a64211feb47e2fa25c048decb3420ee52562fdc4931225c0fb559f4fbb75c378cdd3dd7ccbfeff9c5c2094e5569000000f8678085012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a0921101a4583fb153ec00e501f3c2e2636114e1c8c58d2df8a19426cc066a6768a022db4bce1e0ace485ce0838d178b4d5bcfa9f69b315a14c580d9b01e5c980bdd69000000f8670185012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a06c9842bc78c2f68468cbf8a8f3fec0ae2679707ffb606a4b373dd01a02af55fca01a4c4d984d750678fb204ce8b7e97d860c974ac1a423f098dc4921acd2be0c7d0000000000000000";
@@ -108,60 +46,3 @@ export const serializedProposalDataWithValidRound =
 export const serializedProposalDataWithValidRoundForSignature =
 	"010000000100000000b0020000000174ba618a8c01020000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000200d007000000000000000000000000000000000000000000000000000000000002540be4000000000000000000000000000000000000000000000000000000000000000000e2000000f01a3a2a2990990a64211feb47e2fa25c048decb3420ee52562fdc4931225c0fb559f4fbb75c378cdd3dd7ccbfeff9c5c2094e5569000000f8678085012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a0921101a4583fb153ec00e501f3c2e2636114e1c8c58d2df8a19426cc066a6768a022db4bce1e0ace485ce0838d178b4d5bcfa9f69b315a14c580d9b01e5c980bdd69000000f8670185012a05f200830f424094be89811e15f611c1db12e59679b6f3dc1f4301558080824e43a06c9842bc78c2f68468cbf8a8f3fec0ae2679707ffb606a4b373dd01a02af55fca01a4c4d984d750678fb204ce8b7e97d860c974ac1a423f098dc4921acd2be0c7d000000000000000000";
 
-export const precommitData: Contracts.Crypto.PrecommitData = {
-	blockHash: blockData.hash,
-	blockNumber: 1,
-	round: 1,
-	signature:
-		"8031c22986f2a9e6e8d7d2a90210bb8483e0cad175a35a9a7e3a880f7a383e6af06be29c06199a7f18d4f123bb8a6fd60101f3eabe387f55256b879eb654928c12968f77fdefd0f776dde1806d2865f6a8059f865108aab7e7161f18cf162ee3",
-	type: Enums.Crypto.MessageType.Precommit,
-	validatorIndex: 0,
-};
-
-export const serializedPrecommit =
-	"0201000000010000000182139a7708157c8e2b78f0db38216924c8a17f82e77d5997fb280b1435a6cc97008031c22986f2a9e6e8d7d2a90210bb8483e0cad175a35a9a7e3a880f7a383e6af06be29c06199a7f18d4f123bb8a6fd60101f3eabe387f55256b879eb654928c12968f77fdefd0f776dde1806d2865f6a8059f865108aab7e7161f18cf162ee3";
-
-export const serializedPrecommitForSignature =
-	"0201000000010000000182139a7708157c8e2b78f0db38216924c8a17f82e77d5997fb280b1435a6cc97";
-
-export const precommitDataNoBlock: Contracts.Crypto.PrecommitData = {
-	blockHash: undefined,
-	blockNumber: 1,
-	round: 1,
-	signature:
-		"904c8055242bd7736a1cf7ce20c8fedeee5f2f8fe3f6cab6a166c36c1be0f616c2b7a333912becfa3ecb799c8cd420a012bf41018f5c52f67a2858a6d5bd016e8ef6f56a84d8a734ba6ce5f9a5260201fd9d73ce8688ff0019df2c07a1c33c4d",
-	type: Enums.Crypto.MessageType.Precommit,
-	validatorIndex: 0,
-};
-
-export const serializedPrecommitNoBlock =
-	"0201000000010000000000904c8055242bd7736a1cf7ce20c8fedeee5f2f8fe3f6cab6a166c36c1be0f616c2b7a333912becfa3ecb799c8cd420a012bf41018f5c52f67a2858a6d5bd016e8ef6f56a84d8a734ba6ce5f9a5260201fd9d73ce8688ff0019df2c07a1c33c4d";
-
-export const prevoteData: Contracts.Crypto.PrevoteData = {
-	blockHash: blockData.hash,
-	blockNumber: 1,
-	round: 1,
-	signature:
-		"a25c6d7a0491513cecaabe83f9c1dc59d5d9750a09535fd7d2eb4a81641a425a664c41822c1ba0d414357ad34a8f0a5a084d269f04e8c09da20b44ff75d1a0280919a31d2a519bf0327459471fda97f8a1096500ec84a5b2004eff1dc719de77",
-	type: Enums.Crypto.MessageType.Prevote,
-	validatorIndex: 0,
-};
-
-export const serializedPrevote =
-	"0101000000010000000182139a7708157c8e2b78f0db38216924c8a17f82e77d5997fb280b1435a6cc9700a25c6d7a0491513cecaabe83f9c1dc59d5d9750a09535fd7d2eb4a81641a425a664c41822c1ba0d414357ad34a8f0a5a084d269f04e8c09da20b44ff75d1a0280919a31d2a519bf0327459471fda97f8a1096500ec84a5b2004eff1dc719de77";
-
-export const serializedPrevoteForSignature =
-	"0101000000010000000182139a7708157c8e2b78f0db38216924c8a17f82e77d5997fb280b1435a6cc97";
-
-export const prevoteDataNoBlock: Contracts.Crypto.PrevoteData = {
-	blockHash: undefined,
-	blockNumber: 1,
-	round: 1,
-	signature:
-		"927628d67c385fe216aa800def9cce0c09f5f9fbf836583d7c07ab6a98e1b5681802c92f81ad54984236a07fa389dbab1519f3c91ad39a505a61c3624a88c65da71fe721d7af0ed452516771b94d027be713dba68e14fa2c9680e35b63f0e038",
-	type: Enums.Crypto.MessageType.Prevote,
-	validatorIndex: 0,
-};
-
-export const serializedPrevoteNoBlock =
-	"0101000000010000000000927628d67c385fe216aa800def9cce0c09f5f9fbf836583d7c07ab6a98e1b5681802c92f81ad54984236a07fa389dbab1519f3c91ad39a505a61c3624a88c65da71fe721d7af0ed452516771b94d027be713dba68e14fa2c9680e35b63f0e038";
